Updates to the turning point coordinate differnce for coupled quartic
