+++ src/turning_point_coord_difference.py
@@ -199,7 +199,7 @@ def stateTransitMat(tf,x0,par,varEqns_model,fixed_step=0):
 def turningPoint_configdiff(begin1,begin2, get_coord_model, pot_energy_model, varEqns_model, \
                             configdiff_model, ham2dof_model, half_period_model, \
                             guess_coords_model, plot_iter_orbit_model, par, \
-                            e, n, n_turn, po_fam_file):
+                            e, n, n_turn, show_itrsteps_plots, po_fam_file):
     """
     turningPoint computes the periodic orbit of target energy using turning point based on configuration difference method. 
     
@@ -251,7 +251,10 @@ def turningPoint_configdiff(begin1,begin2, get_coord_model, pot_energy_model, va
         
     n_turn: int
         nth turning point that is used to define the dot product
- 
+
+    show_itrsteps_plots: logical
+        flag (True or False) to show iteration of the UPOs in plots
+
     po_fam_file : function name
         file name to save the members in the family of the unstable periodic orbits      
     Returns
@@ -266,6 +269,7 @@ def turningPoint_configdiff(begin1,begin2, get_coord_model, pot_energy_model, va
         Energy of the target unstable periodic orbit. 
     """
     axis_fs = 15
+    
     #y_PES = get_y(begin1[0], e,par)
     #y_turning = begin1[1]
     #toler = math.sqrt((y_PES-y_turning)**2)
@@ -360,39 +364,17 @@ def turningPoint_configdiff(begin1,begin2, get_coord_model, pot_energy_model, va
                 energy[j] = get_total_energy(x[j,:], pot_energy_model, par)
             energyPO[i_iter]= np.mean(energy)
             
-            ax = plt.gca(projection='3d')
-            plot_iter_orbit_model(x, ax, e, par)
+            if show_itrsteps_plots: # show iteration of the UPOs in plots
+                ax = plt.gca(projection='3d')
+                plot_iter_orbit_model(x, ax, e, par)
             
-#            if model== 'uncoupled':
-#                ax.plot(x[:,0],x[:,1],x[:,3],'-')
-#                ax.plot(x[:,0],x[:,1],-x[:,3],'--')
-#                ax.scatter(x[0,0],x[0,1],x[0,3],s=20,marker='*')
-#                ax.scatter(x[-1,0],x[-1,1],x[-1,3],s=20,marker='o')
-#            elif model== 'coupled':
-#                ax.plot(x[:,0],x[:,1],x[:,3],'-')
-#                ax.plot(x[:,0],x[:,1],-x[:,3],'--')
-#                ax.scatter(x[0,0],x[0,1],x[0,3],s=20,marker='*')
-#                ax.scatter(x[-1,0],x[-1,1],x[-1,3],s=20,marker='o')
-#            elif model == 'deleonberne':
-#                ax.plot(x[:,0],x[:,1],x[:,2],'-')
-#                ax.plot(x[:,0],x[:,1],-x[:,2],'--')
-#                ax.scatter(x[0,0],x[0,1],x[0,2],s=20,marker='*')
-#                ax.scatter(x[-1,0],x[-1,1],x[-1,2],s=20,marker='o')
-#            else:
-#                print("The model you are chosen does not exist ") 
-#                
-#            ax.set_xlabel('$x$', fontsize=axis_fs)
-#            ax.set_ylabel('$y$', fontsize=axis_fs)
-#            ax.set_zlabel('$p_y$', fontsize=axis_fs)
-#            ax.set_title('$\Delta E$ = %e' %(np.mean(energy) - par[2] ), fontsize=axis_fs)
-#            ax.set_xlim(-1, 1)
-#            ax.set_ylim(-1, 1)
-            #x_turn= x[-1,0]  # x coordinate of turning point
-            #y_turn= x[-1,1] # y coordinate of turning point
-            #y_PES = -get_y(x_turn,e,par)
-            #toler = math.sqrt((y_PES-y_turn)**2)
-            plt.grid()
-            plt.show()
+                #x_turn= x[-1,0]  # x coordinate of turning point
+                #y_turn= x[-1,1] # y coordinate of turning point
+                #y_PES = -get_y(x_turn,e,par)
+                #toler = math.sqrt((y_PES-y_turn)**2)
+                plt.grid()
+                plt.show()
+
             guess2 = np.array([result[index,1], result[index,2],0,0])
             guess1 = np.array([result[index-1,1], result[index-1,2],0,0])
             
@@ -447,254 +429,5 @@ def turningPoint_configdiff(begin1,begin2, get_coord_model, pot_energy_model, va
 
     
 
-#%% get potential energy
-#def get_potential_energy(model,x,y,par):
-#    """ enter the definition of the potential energy function  
-#    """      
-#    
-#    if model == 'uncoupled':
-#        pot_energy =  -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2
-#    elif model == 'coupled':
-#        pot_energy =  -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2+0.5*par[6]*(x-y)**2
-#    elif model== 'deleonberne':
-#        pot_energy = par[3]*( 1 - math.e**(-par[4]*x) )**2 + 4*y**2*(y**2 - 1)*math.e**(-par[5]*par[4]*x) + par[2]
-#    else:
-#        print("The model you are chosen does not exist, enter the definition of the potential ernergy")
-#                
-#    return pot_energy
-
-#%
-#def func_vec_field_eq_pt(model,x,par):
-#    """ vecor field(same as pxdot, pydot), used to find the equilibrium points of the system.
-#    """
-#    if model == 'uncoupled':
-#        dVdx = -par[3]*x[0]+par[4]*(x[0])**3
-#        dVdy = par[5]*x[1]
-#    elif model == 'coupled':
-#        dVdx = (-par[3]+par[6])*x[0]+par[4]*(x[0])**3-par[6]*x[1]
-#        dVdy = (par[5]+par[6])*x[1]-par[6]*x[0]
-#    elif model == 'deleonberne':
-#        dVdx = -2*par[3]*par[4]*math.e**(-par[4]*x[0])*(math.e**(-par[4]*x[0]) - 1) - 4*par[5]*par[4]*x[1]**2*(x[1]**2 - 1)*math.e**(-par[5]*par[4]*x[0])
-#        dVdy = 8*x[1]*(2*x[1]**2 - 1)*math.e**(-par[5]*par[4]*x[0])
-#    else:
-#        print("The model you are chosen does not exist, enter the definition of Hamilton's equations")
-#    F = [-dVdx, -dVdy]
-#    return F
-
-
-#%
-#def get_eq_pts(eqNum,model, par):
-#    """
-#    GET_EQ_PTS_BP solves the saddle center equilibrium point for a system with
-#    KE + PE. 
-#     H = 1/2*px^2+omega/2*py^2 -(1/2)*alpha*x^2+1/4*beta*x^4+ omega/2y^2 with alpha > 0
-#    --------------------------------------------------------------------------
-#       Uncouled potential energy surface notations:
-#    
-#               Well (stable, EQNUM = 2)    
-#    
-#                   Saddle (EQNUM=1)
-#    
-#               Well (stable, EQNUM = 3)    
-#    
-#    --------------------------------------------------------------------------
-#       
-#    
-#    """
-#    #fix the equilibrium point numbering convention here and make a
-#    #starting guess at the solution
-#    if 	eqNum == 1:
-#        
-#        x0 = [0, 0]                  # EQNUM = 1, saddle  
-#    elif 	eqNum == 2: 
-#        if model == 'uncoupled':
-#            eqPt = [+math.sqrt(par[3]/par[4]),0]    # EQNUM = 2, stable
-#            return eqPt
-#        elif model == 'coupled':
-#            eqPt = [+math.sqrt(par[3]-par[6]/par[4]),0] # EQNUM = 2, stable
-#            return eqPt
-#        elif model== 'deleonberne':
-#            eqPt = [0, 1/math.sqrt(2)]    # EQNUM = 2, stable
-#            return eqPt
-#        else:
-#            print("The model you are chosen does not exist, enter the definition of Hamilton's equations")
-#    elif 	eqNum == 3:
-#        if model == 'uncoupled':
-#            eqPt = [-math.sqrt(par[3]/par[4]),0]    # EQNUM = 2, stable
-#            return eqPt
-#        elif model == 'coupled':
-#            eqPt = [-math.sqrt(par[3]-par[6]/par[4]),0] # EQNUM = 2, stable
-#            return eqPt
-#        elif model== 'deleonberne':
-#            eqPt = [0, -1/math.sqrt(2)]    # EQNUM = 2, stable
-#            return eqPt
-#        else:
-#            print("The model you are chosen does not exist, enter the definition of Hamilton's equations")
-#    
-#    # F(xEq) = 0 at the equilibrium point, solve using in-built function
-#    F = lambda x: func_vec_field_eq_pt(model,x,par)
-#    eqPt = fsolve(F,x0, fprime=None) # Call solver
-#    return eqPt
-
-
-#def stateTransitMat(tf,x0,par,model,fixed_step=0): 
-#    """
-#    function [x,t,phi_tf,PHI] =
-#    stateTransitionMatrix_boatPR(x0,tf,R,OPTIONS,fixed_step)
-#    
-#    Gets state transition matrix, phi_tf = PHI(0,tf), and the trajectory 
-#    (x,t) for a length of time, tf.  
-#    
-#    In particular, for periodic solutions of % period tf=T, one can obtain 
-#    the monodromy matrix, PHI(0,T).
-#    """
-#    
-#
-#    N = len(x0)  # N=4 
-#    RelTol=3e-14
-#    AbsTol=1e-14  
-#    tf = tf[-1]
-#    if fixed_step == 0:
-#        TSPAN = [ 0 , tf ] 
-#    else:
-#        TSPAN = np.linspace(0, tf, fixed_step)
-#    PHI_0 = np.zeros(N+N**2)
-#    PHI_0[0:N**2] = np.reshape(np.identity(N),(N**2)) # initial condition for state transition matrix
-#    PHI_0[N**2:N+N**2] = x0                    # initial condition for trajectory
-#
-#    
-#    f = lambda t,PHI: varEqns(t,PHI,par, model) # Use partial in order to pass parameters to function
-#    soln = solve_ivp(f, TSPAN, list(PHI_0),method='RK45',dense_output=True, events = None,rtol=RelTol, atol=AbsTol)
-#    t = soln.t
-#    PHI = soln.y
-#    PHI = PHI.transpose()
-#    x = PHI[:,N**2:N+N**2]		   # trajectory from time 0 to tf
-#    phi_tf = np.reshape(PHI[len(t)-1,0:N**2],(N,N)) # state transition matrix, PHI(O,tf)
-#
-#    
-#    return t,x,phi_tf,PHI
-
-
-#%
-#def varEqns(t,PHI,par,model):
-#    """
-#    PHIdot = varEqns_bp(t,PHI) 
-#    
-#    This here is a preliminary state transition, PHI(t,t0),
-#    matrix equation attempt for a ball rolling on the surface, based on...
-#    
-#    d PHI(t, t0)
-#    ------------ =  Df(t) * PHI(t, t0)
-#        dt
-#    
-#    """
-#    phi = PHI[0:16]
-#    phimatrix  = np.reshape(PHI[0:16],(4,4))
-#    x,y,px,py = PHI[16:20]
-#    
-#    
-#    if model == 'uncoupled':
-#        # The first order derivative of the Hamiltonian.
-#        dVdx = -par[3]*x+par[4]*x**3
-#        dVdy = par[5]*y
-#    
-#        # The following is the Jacobian matrix 
-#        d2Vdx2 = -par[3]+par[4]*3*x**2
-#            
-#        d2Vdy2 = par[5]
-#    
-#        d2Vdydx = 0
-#    elif model == 'coupled':
-#        # The first order derivative of the Hamiltonian.
-#        dVdx = (-par[3]+par[6])*x+par[4]*x**3-par[6]*y
-#        dVdy = (par[5]+par[6])*y-par[6]*x
-#
-#        # The following is the Jacobian matrix 
-#        d2Vdx2 = -par[3]+par[6]+par[4]*3*x**2
-#            
-#        d2Vdy2 = par[5]+par[6]
-#    
-#        d2Vdydx = -par[6]
-#    elif model == 'deleonberne':
-#        # The first order derivative of the Hamiltonian.
-#        dVdx = - 2*par[3]*par[4]*math.e**(-par[4]*x)*(math.e**(-par[4]*x) - 1) - 4*par[5]*par[4]*y**2*(y**2 - 1)*math.e**(-par[5]*par[4]*x) 
-#        dVdy = 8*y*(2*y**2 - 1)*math.e**(-par[5]*par[4]*x)
-#    
-#        # The following is the Jacobian matrix 
-#        d2Vdx2 = - ( 2*par[3]*par[4]**2*( math.e**(-par[4]*x) - 2.0*math.e**(-2*par[4]*x) ) - 4*(par[5]*par[4])**2*x**2*(y**2 - 1)*math.e**(-par[5]*par[4]*x) )
-#            
-#        d2Vdy2 = 8*(6*y**2 - 1)*math.e**( -par[4]*par[5]*x )
-#    
-#        d2Vdydx = -8*y*par[4]*par[5]*math.e**( -par[4]*par[5]*x )*(2*y**2 - 1)
-#        
-#    else:
-#        print("The model you are chosen does not exist")
-#    
-#    
-#    d2Vdxdy = d2Vdydx    
-#
-#    Df    = np.array([[  0,     0,    par[0],    0],
-#              [0,     0,    0,    par[1]],
-#              [-d2Vdx2,  -d2Vdydx,   0,    0],
-#              [-d2Vdxdy, -d2Vdy2,    0,    0]])
-#
-#    
-#    phidot = np.matmul(Df, phimatrix) # variational equation
-#
-#    PHIdot        = np.zeros(20)
-#    PHIdot[0:16]  = np.reshape(phidot,(1,16)) 
-#    PHIdot[16]    = px/par[0]
-#    PHIdot[17]    = py/par[1]
-#    PHIdot[18]    = -dVdx 
-#    PHIdot[19]    = -dVdy
-#    
-#    return list(PHIdot)
-
-#%
-#def half_period(t,x,model):
-#    """
-#    Return the turning point where we want to stop the integration                           
-#    
-#    pxDot = x[0]
-#    pyDot = x[1]
-#    xDot = x[2]
-#    yDot = x[3]
-#    """
-#    terminal = True
-#    # The zero can be approached from either direction
-#    direction = 0 #0: all directions of crossing
-#    if model =='uncoupled':
-#        return x[3]
-#    elif model =='coupled':
-#        return x[3]
-#    elif model =='deleonberne':
-#        return x[2]
-#    else:
-#        print("specify an event for integration, either choose to return p_x or to return p_y")
-
-
-#% Hamilton's equations
-#def Ham2dof(model,t,x, par):
-#    """ Enter the definition of Hamilton's equations, used for ploting trajectories later.
-#    """
-#    xDot = np.zeros(4)
-#    
-#    if model == 'uncoupled':
-#        dVdx = -par[3]*x[0]+par[4]*(x[0])**3
-#        dVdy = par[5]*x[1]
-#    elif model == 'coupled':
-#        dVdx = (-par[3]+par[6])*x[0]+par[4]*(x[0])**3-par[6]*x[1]
-#        dVdy = (par[5]+par[6])*x[1]-par[6]*x[0]
-#    elif model == 'deleonberne':
-#        dVdx = -2*par[3]*par[4]*math.e**(-par[4]*x[0])*(math.e**(-par[4]*x[0]) - 1) - 4*par[5]*par[4]*x[1]**2*(x[1]**2 - 1)*math.e**(-par[5]*par[4]*x[0])
-#        dVdy = 8*x[1]*(2*x[1]**2 - 1)*math.e**(-par[5]*par[4]*x[0])
-#    else:
-#        print("The model you are chosen does not exist, enter the definition of Hamilton's equations")
-#        
-#    xDot[0] = x[2]/par[0]
-#    xDot[1] = x[3]/par[1]
-#    xDot[2] = -dVdx 
-#    xDot[3] = -dVdy
-#    return list(xDot)
